refactor(sampler): removed renderCell and moved type enums
